Remove lodash-es dependency from client lib
It turns out that under some consumer configurations the lodash-es dep isn't able to load correctly. We aren't using it for much so I've just removed it entirely.

This PR also updates our OpenAI client library version. It turns out that this wasn't actually necessary, but it's probably good to stay on the latest version anyway.