Better type safety for inputs and db operations
All the casting `as unknown as ...` makes it really easy to lose track of what types are in each column and received by each endpoint. This PR adds new Zod validators for the JSON columns we work with most frequently and tries to apply them whenever we read from the database so we get good types, and will get alerted if a wrongly-typed value makes it into the database in prod.

Also updates our `/chat/completions` endpoint to be compatible with OpenAI.