Count dataset entry tokens in the background
Previously we were counting dtaset entry tokens as we were importing the entries. This didn't work well for large datasets -- in fact I was getting timeouts when trying to construct a dataset of size 10k.

This PR moves that calculation to the background. It also sprinkles in a number of small formatting/text changes.