Speed up the request logs page
Did some postgres query analysis and fixed the slowest parts of the request logs page.

It's still slow for users that legitimately have a lot of logs because of the query to get the unique tag names. I think the only solution there will be to change our database structure so we keep a list of all tags that have been applied in each project, but didn't want to block this PR on that change. This will improve performance significantly for users who do not have a large number of logs.