Add response_format to DatasetEntries
This way if a user requests JSON, we (1) pass that through to the gpt-3.5 comparison on the evals page, and (2) automatically run our accuracy checks as if it were a function call.

I didn't make the changes to format the output as JSON and show accuracy, since I think David will be touching that code imminently anyway.