Fix server crash caused by isolated VM package
Pass --no-node-snapshot in NODE_OPTIONS to the next dev command. This is required according to the maintainers of isolated VM https://github.com/laverdet/isolated-vm/issues/420. This was causing a very difficult to debug issue where any trpc requests would return a 500 internal server error with this unhelpful error message. I was able to track it down to this cause by commenting out sections of server code until I found it.

```sh
[next] - error Error: socket hang up
[next]     at connResetException (node:internal/errors:721:14)
[next]     at Socket.socketOnEnd (node:_http_client:519:23)
[next]     at Socket.emit (node:events:526:35)
[next]     at Socket.emit (node:domain:488:12)
[next]     at endReadableNT (node:internal/streams/readable:1408:12)
[next]     at process.processTicksAndRejections (node:internal/process/task_queues:82:21) {
[next]   code: 'ECONNRESET'
[next] }
```

Not sure if it's because I'm running on an M2 Mac? But adding the node option fixes it.