Add "Results" tab to evals
A couple of simple visualizations: a table view as well as a heatmap of head to head results.

As I worked on this I became more convinced that we should split this into a new "Evaluations" tab and rename the current tab to "Test Set" or something like that. For one thing, it's important that we give people a way to link to and share evaluations! But didn't want to take the time to do that now.