Don't try to load all the training data into memory at once
This implements a new training data upload pattern that grabs 1000 rows from the database at a time, processes them, and then uploads them directly to Azure. That way we never have to have all of our training data in memory at once.

This is likely a pattern we'll have to implement in more places over time as datasets get larger.