Set and track training configs app-side
This change makes our TS app responsible for generating the full Axolotl config for a training run, instead of the Python trainer. There are two reasons for this:
 - It lets us more easily experiment with different training configs in development.
 - It will let us allow end-users to more easily override specific hparams (eg. number of epochs, learning rate, etc.) in the future.

I'm working on this right now because we need to test different hparams as part of figuring out the right balance for LoRA hot-swapping.