Use S-LoRA for inference
This PR switches us to the S-LoRA branch of vLLM and starts serving future fine-tunes via S-LoRA, instead of by serving the fully-merged model.

It does however maintain backwards compatibility. After training a model, we still persist the merged model to disk, and we could switch to serving it via the old process just by setting the fine-tune's pipeline version to 2 instead of 3.

There is an (intentional) regression here: for new pipeline v3 models trained in prod, we'll no longer be able to automatically serve them in development. I'll need to write a script that we can call to explicitly copy production models into development when necessary for testing.