Use more routing functions in Modal
It turns out that somewhere in Modal's stack there's a bottleneck where sending too many simultaneous requests to the same function slows down, even if the function itself is very lightweight. This is a workaround that spins up a new instance of our routing function for every 20 requests to avoid the bottleneck.