Reload the volume if we can't find a model's weights
It turns out you have to explicitly reload a volume once you've made changes to it. So long-lived inference workers weren't seeing the weights from models trained after they started.

This also slows down the retry schedule for inference so it runs over several days, not just 12 minutes. It's more likely that we'll fix whatever issue is causing the completions to fail over a longer horizon.