Manage cursor position in fine-tune name box
Previously, the cursor would always jump to the end of the input box
when the user tried to enter an invalid character in the fine-tune name. This commit ensures that the cursor stays in the expected position when the user enters an invalid character that is replaced by a hyphen.