Fireworks rate limiting
Ensure we never send more than 30 requests to Fireworks as part of evals at once, and that we send requests at a maximum rate of 10 rps.

Also adds a general `rateLimit` abstraction that we may be able to reuse for similar types of soft limits in the future. It uses the token bucket strategy -- leaky bucket would probably be a better fit semantically, but would require some pretty major changes to the way we queue/process background jobs so decided to just go with the simple approach instead.